+++ src/CommerceMolliePlusPlugin.php
@@ -5,6 +5,7 @@ namespace white\commerce\mollie\plus;
 use craft\base\Plugin;
 use craft\commerce\elements\Order;
 use craft\commerce\Plugin as CommercePlugin;
+use craft\commerce\records\Transaction;
 use craft\commerce\records\Transaction as TransactionRecord;
 use craft\commerce\services\Gateways;
 use craft\events\ModelEvent;
@@ -23,9 +24,13 @@ class CommerceMolliePlusPlugin extends Plugin
     {
         parent::init();
 
-        Event::on(Gateways::class, Gateways::EVENT_REGISTER_GATEWAY_TYPES,  function(RegisterComponentTypesEvent $event) {
-            $event->types[] = Gateway::class;
-        });
+        Event::on(
+            Gateways::class,
+            Gateways::EVENT_REGISTER_GATEWAY_TYPES,
+            function(RegisterComponentTypesEvent $event) {
+                $event->types[] = Gateway::class;
+            }
+        );
 
         $this->registerOrderEventListeners();
     }
@@ -51,6 +56,26 @@ class CommerceMolliePlusPlugin extends Plugin
                             if ($child->status == TransactionRecord::STATUS_SUCCESS) {
                                 $child->order->updateOrderPaidInformation();
                             }
+                        }
+                    }
+                }
+            }
+        );
+
+        Event::on(
+            Order::class,
+            Order::EVENT_BEFORE_COMPLETE_ORDER,
+            function(Event $event) {
+                /** @var Order $order */
+                $order = $event->sender;
+
+                $transaction = $order->getLastTransaction();
+                $gateway = $transaction->getGateway() ?? null;
+                if ($gateway instanceof Gateway && !$gateway->completeBanktransferOrders) {
+                    if ($transaction->status === Transaction::STATUS_PROCESSING) {
+                        $transactionMessage = json_decode($transaction->message);
+                        if ($transactionMessage->method === 'banktransfer') {
+                            $order->isCompleted = false;
                         }
                     }
                 }